+++ src/commands/hardis/mdapi/deploy.ts
@@ -8,7 +8,7 @@ const xorFlags = ["zipfile", "validateddeployrequestid", "deploydir"];
 export class Deploy extends SfdxCommand {
   public static readonly description = `sfdx-hardis wrapper for sfdx force:mdapi:deploy that displays tips to solve deployment errors.
 
-See documentation at https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_force_mdapi.htm#cli_reference_force_mdapi_deploy
+[See documentation of Salesforce command](https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_force_mdapi.htm#cli_reference_force_mdapi_deploy)
 `;
   public static readonly examples = [];
   public static readonly requiresUsername = true;