+++ src/commands/hardis/project/deploy/sources/dx.ts
@@ -3,9 +3,10 @@
 import { flags, SfdxCommand } from "@salesforce/command";
 import { Messages } from "@salesforce/core";
 import { AnyJson } from "@salesforce/ts-types";
+import * as c from "chalk";
 import * as fs from "fs-extra";
 import { MetadataUtils } from "../../../../../common/metadata-utils";
-import { isCI } from "../../../../../common/utils";
+import { isCI, uxLog } from "../../../../../common/utils";
 import { getConfig } from "../../../../../config";
 import { forceSourceDeploy } from "../../../../../common/utils/deployUtils";
 import { promptOrg } from "../../../../../common/utils/orgUtils";
@@ -30,7 +31,7 @@ If necessary,You can define the following files (that supports wildcards <member
 Env vars override:
 
 - SFDX_HARDIS_DEPLOY_IGNORE_SPLIT_PACKAGES: define "true" to ignore split of package.xml into several deployments
-
+- INSTALL_PACKAGES_DURING_CHECK_DEPLOY: define "true" so packages are installed during deployment check
   `;
 
   public static examples = ["$ sfdx hardis:project:deploy:sources:dx"];
@@ -79,17 +80,33 @@ Env vars override:
     const packageXml = this.flags.packagexml || null;
     this.debugMode = this.flags.debug || false;
 
-    // Install packages
-    const packages = this.configInfo.installedPackages || [];
-    if (packages.length > 0 && !check) {
-      // Install package only if we are in real deployment mode
-      await MetadataUtils.installPackagesOnOrg(packages, null, this, "deploy");
-    }
-
+    // Get target org
     let targetUsername = this.org.getUsername();
     if (!isCI) {
       const targetOrg = await promptOrg(this, { devHub: false, setDefault: false, scratch: false });
       targetUsername = targetOrg.username;
+    }
+
+    // Install packages
+    const packages = this.configInfo.installedPackages || [];
+    const missingPackages = [];
+    const installPackages =
+      check === false || process.env.INSTALL_PACKAGES_DURING_CHECK_DEPLOY === "true" || this.configInfo.installPackagesDuringCheckDeploy === true;
+    if (packages.length > 0 && installPackages) {
+      // Install packages only if we are in real deployment mode
+      await MetadataUtils.installPackagesOnOrg(packages, targetUsername, this, "deploy");
+    } else if (packages.length > 0 && check === true) {
+      // If check mode, warn if there are missing packages
+      const alreadyInstalled = await MetadataUtils.listInstalledPackages(targetUsername, this);
+      for (const package1 of packages) {
+        if (
+          alreadyInstalled.filter((installedPackage: any) => package1.SubscriberPackageVersionId === installedPackage.SubscriberPackageVersionId)
+            .length === 0 &&
+          package1.installDuringDeployments === true
+        ) {
+          missingPackages.push(package1);
+        }
+      }
     }
 
     // Get package.xml
@@ -109,6 +126,32 @@ Env vars override:
     if (fs.existsSync(packageDeletedXmlFile)) {
       forceSourceDeployOptions.postDestructiveChanges = packageDeletedXmlFile;
     }
+
+    // Display missing packages message
+    if (missingPackages.length > 0) {
+      for (const package1 of missingPackages) {
+        uxLog(
+          this,
+          c.yellow(
+            `You may need to install package ${c.bold(package1.SubscriberPackageName)} ${c.bold(
+              package1.SubscriberPackageVersionId
+            )} in target org to validate the deployment check`
+          )
+        );
+      }
+      uxLog(this, "");
+      uxLog(
+        this,
+        c.yellow(
+          c.italic(
+            `If you want deployment checks to automatically install packages, please define ${c.bold(
+              "INSTALL_PACKAGES_DURING_CHECK_DEPLOY=true"
+            )} in ENV vars, or property ${c.bold("installPackagesDuringCheckDeploy: true")} in .sfdx-hardis.yml`
+          )
+        )
+      );
+    }
+
     // Process deployment (or deployment check)
     const { messages } = await forceSourceDeploy(packageXmlFile, check, testlevel, this.debugMode, this, forceSourceDeployOptions);